:art: Comment 인가 처리 로직 수정

+++ src/main/kotlin/team/sparta/onehouronemeal/domain/comment/service/v1/CommentService.kt
@@ -1,9 +1,10 @@
 package team.sparta.onehouronemeal.domain.comment.service.v1
 
 import org.springframework.data.repository.findByIdOrNull
+import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-import team.sparta.onehouronemeal.domain.auth.common.PermissionChecker
+import org.springframework.web.server.ResponseStatusException
 import team.sparta.onehouronemeal.domain.comment.dto.v1.CommentResponse
 import team.sparta.onehouronemeal.domain.comment.dto.v1.CreateCommentRequest
 import team.sparta.onehouronemeal.domain.comment.dto.v1.UpdateCommentRequest
@@ -43,7 +44,7 @@ class CommentService(
     fun updateComment(principal: UserPrincipal, commentId: Long, request: UpdateCommentRequest): CommentResponse {
         val comment = commentRepository.findByIdOrNull(commentId) ?: throw ModelNotFoundException("Comment", commentId)
 
-        PermissionChecker.check(comment.user.id!!, principal)
+        checkPermission(comment, principal)
 
         comment.updateComment(request.content)
 
@@ -54,7 +55,7 @@ class CommentService(
     fun deleteComment(principal: UserPrincipal, commentId: Long) {
         val comment = commentRepository.findByIdOrNull(commentId) ?: throw ModelNotFoundException("Comment", commentId)
 
-        PermissionChecker.check(comment.user.id!!, principal)
+        checkPermission(comment, principal)
 
         commentRepository.delete(comment)
     }
@@ -67,5 +68,14 @@ class CommentService(
         // 본인 댓글에 신고 안됨
 
         TODO()
+    }
+
+    private fun checkPermission(comment: Comment, principal: UserPrincipal) {
+        check(
+            comment.checkPermission(
+                principal.id,
+                principal.role
+            )
+        ) { throw ResponseStatusException(HttpStatus.FORBIDDEN, "Permission denied") }
     }
 }
